requirements.txt
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.28.1
email-validator==1.1.2
fastapi==0.62.0
PyJWT==1.7.1
python-decouple==3.3
uvicorn==0.12.3